// server/src/socket/table/table.types.ts
export enum ETableEvents {
    ADD_ROW = 'ADD_ROW',
    ROW_ADDED = 'ROW_ADDED',
    REMOVE_ROW = 'REMOVE_ROW',
    ROW_REMOVED = 'ROW_REMOVED',
    EDIT_ROW = 'EDIT_ROW',
    ROW_EDITED = 'ROW_EDITED',
    EDIT_COLUMN = 'EDIT_COLUMN',
    COLUMN_EDITED = 'COLUMN_EDITED',
    ADD_COLUMN = 'ADD_COLUMN',
    COLUMN_ADDED = 'COLUMN_ADDED',
    REMOVE_COLUMN = 'REMOVE_COLUMN',
    COLUMN_REMOVED = 'COLUMN_REMOVED'
}